More checks (dig)
Add dig checks for cloudflare and google. This is used for checking the
own hostname on their nameservers as well. In special this is useful
when the server is behind a dynamic dns network where things with DNS
could go wrong.
Also it is allowed to test more than one record with the normal dig check.
The things done in this commit are a bit ugly, but fortunately @n0emis
said they would try to make it better than me (@n0emis can do things much
better than me in general anyway)